some little changes (#6)
* cosmetic changes

* download wget before pirl installation

* download wget before pirl installation

* install which

* download wget before pirl installation

* ufw allow new ports for v5

+++ firewall_installer.sh
@@ -101,11 +101,13 @@ done
 
 echo "Updating/Installing packages.  This will take a few minutes."
 
-#install which because some distributives can come without it
+# install which because some distributives can come without it
 
-apt install which -y >/dev/null 2>/dev/null
-apt-get install which -y >/dev/null 2>/dev/null
-yum install -y which >/dev/null 2>/dev/null
+if [ ! -f /usr/bin/which ] ; then
+	apt install which -y >/dev/null 2>/dev/null
+	apt-get install which -y >/dev/null 2>/dev/null
+	yum install -y which >/dev/null 2>/dev/null
+fi
 
 ############## update packages ###################
 #determine if apt, apt-get or yum
@@ -229,6 +231,9 @@ fi
 
 #default ports for pirlnode
 ufw allow 30303
+ufw allow 4001
+ufw allow 6588
+ufw allow 6589
 
 #allow all outgoing
 ufw default allow outgoing
